feat: refactor + readme.md update
