+++ app/src/middleware/validateSession.js
